Divided code into components
