+++ code/src/components/RadioForm.js
@@ -1,0 +1,31 @@
+import React from 'react'
+
+const RadioForm = ({roles, setRoles}) => {
+    const rolesList = [
+        'Frontend',
+        'Backend',
+        'Fullstack',
+        'Cloud'
+    ]
+
+    return (
+        <div className="radio-form-div">
+            <div>
+                <h2>Interested in working with (select one)</h2>
+                {rolesList.map(role => (
+                <label key={role}>
+                <input 
+                    type="radio"
+                    value={role}
+                    onChange={event => setRoles(event.target.value)}
+                    checked={false ? roles : roles === role}                      
+                />
+                {role}
+                </label>   
+                ))}
+            </div> 
+        </div>
+    )
+}
+
+export default RadioForm;
